+++ R/duration.R
@@ -10,6 +10,12 @@
 #' @param grouping.var The grouping variables.  Default \code{NULL} generates
 #' one word list for all text.  Also takes a single grouping variable or a list
 #' of 1 or more grouping variables.
+#' @param as.tibble logical.  If \code{TRUE} the output class will be set to a
+#' \pkg{tibble}, otherwise a \code{\link[data.table]{data.table}}.  Default
+#' checks \code{getOption("tibble.out")} as a logical.  If this is \code{NULL}
+#' the default \code{\link[textshape]{tibble_output}} will set \code{as.tibble}
+#' to \code{TRUE} if \pkg{dplyr} is loaded.  Otherwise, the output will be a
+#' \code{\link[data.table]{data.table}}.
 #' @param \ldots Ignored.
 #' @export
 #' @rdname duration
@@ -90,7 +96,7 @@ duration.default <- function(x, grouping.var = NULL, ...) {
 #' @export
 #' @rdname duration
 #' @method duration data.frame
-duration.data.frame <- function(x, text.var = TRUE, ...) {
+duration.data.frame <- function(x, text.var = TRUE, as.tibble = tibble_output(), ...) {
 
     word.count <- NULL
     nms <- colnames(x)
@@ -113,7 +119,7 @@ duration.data.frame <- function(x, text.var = TRUE, ...) {
 
     colord <- c(nms[!nms %in% text.var], "word.count", "start", "end", text.var)
     data.table:: setcolorder(z, colord)
-    z[]
+    if_tibble(z[], as.tibble = as.tibble)
 }