+++ R/tidy_colo_dtm.R
@@ -5,12 +5,20 @@
 #' a tidy data set made of collocating words.
 #'
 #' @param x A \code{\link[tm]{DocumentTermMatrix}}/\code{\link[tm]{TermDocumentMatrix}}.
-#' @param \ldots ignored.
+#' @param as.tibble logical.  If \code{TRUE} the output class will be set to a
+#' \pkg{tibble}, otherwise a \code{\link[data.table]{data.table}}.  Default
+#' checks \code{getOption("tibble.out")} as a logical.  If this is \code{NULL}
+#' the default \code{\link[textshape]{tibble_output}} will set \code{as.tibble}
+#' to \code{TRUE} if \pkg{dplyr} is loaded.  Otherwise, the output will be a
+#' \code{\link[data.table]{data.table}}.
+#' @param \ldots Ignored.
 #' @return Returns a tidied data.frame.
 #' @rdname tidy_colo_dtm
 #' @export
 #' @seealso \code{\link[textshape]{unique_pairs}}
 #' @examples
+#' data(simple_dtm)
+#'
 #' tidied <- tidy_colo_dtm(simple_dtm)
 #' tidied
 #' unique_pairs(tidied)
@@ -34,34 +42,40 @@
 #'         scale_fill_gradient(low= 'white', high = 'red') +
 #'         theme(axis.text.x = element_text(angle = 45, hjust = 1))
 #' }
-tidy_colo_tdm <- function(x, ...){
+tidy_colo_tdm <- function(x, as.tibble = tibble_output(), ...){
 
     term_1 <- NULL
 
     x <- slam::as.simple_triplet_matrix(slam::tcrossprod_simple_triplet_matrix(x, y = NULL))
 
-    data.table::data.table(
-        term_1 = x[['dimnames']][['Terms']][x[['i']]],
-        term_2 = x[['dimnames']][['Terms']][x[['j']]],
-        n = x[['v']]
-    )[order(term_1), ]
+    if_tibble(
+        data.table::data.table(
+            term_1 = x[['dimnames']][['Terms']][x[['i']]],
+            term_2 = x[['dimnames']][['Terms']][x[['j']]],
+            n = x[['v']]
+        )[order(term_1), ],
+        as.tibble = as.tibble
+    )
 }
 
 
 
 #' @rdname tidy_colo_dtm
 #' @export
-tidy_colo_dtm <- function(x, ...){
+tidy_colo_dtm <- function(x, as.tibble = tibble_output(), ...){
 
     term_1 <- NULL
 
     x <- slam::as.simple_triplet_matrix(slam::crossprod_simple_triplet_matrix(x, y = NULL))
 
-    data.table::data.table(
-        term_1 = x[['dimnames']][['Terms']][x[['i']]],
-        term_2 = x[['dimnames']][['Terms']][x[['j']]],
-        n = x[['v']]
-    )[order(term_1), ]
+    if_tibble(
+        data.table::data.table(
+            term_1 = x[['dimnames']][['Terms']][x[['i']]],
+            term_2 = x[['dimnames']][['Terms']][x[['j']]],
+            n = x[['v']]
+        )[order(term_1), ],
+        as.tibble = as.tibble
+    )
 }